refactor: unpublish vo-test-harness
