+++ packages/swingset-liveslots/test/vo-test-harness.test.js
@@ -1,5 +1,5 @@
 import test from 'ava';
-import { makeSpy, runVOTest } from '../tools/vo-test-harness.js';
+import { makeSpy, runVOTest } from './vo-test-harness.js';
 
 async function voTestTest(t, mode) {
   let makeThing;
